Development ultragraphql 1.1.7

+++ src/model/orderedQuestion.js
@@ -26,12 +26,12 @@ export const orderedQuestion = {
         points: {
             required: true,
             multiple: false,
-            dataType: "integer",
+            dataType: "decimal",
         },
         score: {
             required: false,
             multiple: false,
-            dataType: "integer",
+            dataType: "decimal",
         },
         comment: {
             required: false,
